Create Inventory.pas
When the total inventory cost C in a large store is given by C = (Csq/2)+(Cdr/q) where Cs is the unit storage cost per unit time; Cd  is the cost delivery; q is the number of times in each delivery (the lot size); and r is the rate at which the item is sold (the demand rate). The optimal lot-size qo (i.e. the lot-size that minimizes the total cost) is given by qo = sqrt(2Cdr/C8). Assuming that the unit storage cost per week and delivery cost are constant, write a program to read in the current lot-size and demand rate and to calculate (a) the current inventory cost; (b) the optimal inventory cost; (c) the saving if the optimal lot-size is used.