Create BankAccount.pas
This program processes deposits and withdrawals for a bank account. The input consists of the initial balance followed by a set of deposits and withdrawals, each represented by the letter D or W followed by an amount. The output consists of a statement with the heading ‘DEPOSIT WITHDRAWAL CREDIT DEBIT’ where an entry is made in the third or fourth column, depending upon whether or not the account is overdrawn