Create PseudoRandomNumbers.pas
If your Pascal implementation has a function for generating pseudo-random numbers then you may use this to obtain an approximation to π. The procedure involves generating a random point in a unit square (by generating two random numbers between 0 and 1), and determining whether this point is inside or outside the circle of diameter 1 with centre at (1/2, 1/2). This process is repeated for a large number of pints (5000, say). The proportion of ‘hits’ (i.e. points lying inside or on the circle) to the total number of attempts will be roughly equal to the ratio of the area of the circle to that of the square (i.e. π/4). Use this technique in a program for approximating π. See how accurate the result is for various number of attempts.