Create LargestNumber.pas
Here's a Pascal program that will output the largest number of three integer values.  This program first prompts the user to enter three integer values, which are stored in the variables num1, num2, and num3. Then, it uses a series of conditional statements (if-else) to compare these three values and determine which one is the largest. The largest value is stored in the variable largest, and then the program outputs this value to the console