Create StaticArray.pas
A static array has a fixed size, which is specified when the array is declared. This program is an example of a static array. 
In this example, we declare a static array myArray of size 5 using the syntax array[1..5] of integer. 
We assign values to the array elements individually and then print them out using a for loop.