Create DynamicArray.pas
A dynamic array has a flexible size, which can be changed at runtime.
In this example, we declare a dynamic array myArray using the syntax array of integer. We set the length of the array to 5 using the SetLength function, and then assign values to the array elements individually. Finally, we print out the array elements using a for loop