Create OpenArray.pas
An open array is a special type of parameter that allows the function or procedure to accept arrays of any size.
In this program, we declare a dynamic array myArray using the syntax array of integer. We define a procedure PrintArray that accepts an open array parameter of type array of integer. The High function is used to get the highest index of the array. We then assign values to myArray and pass it to the PrintArray procedure, which prints out