Create FindHighest.pas
This program accepts three integer numbers, determines the highest among the three using a procedure, and outputs the result on the screen.

The DetermineHighest procedure takes four arguments: x, y, z, and highest. It uses nested if statements to determine which of the three integers is the highest and then sets the highest variable to that value.

In the main program, the user is prompted to enter three integers, which are then passed as arguments to the DetermineHighest procedure along with the highest variable. The resulting highest value is then output to the console