Create ExamResults.pas
In a certain examination scheme, four papers are set. These are paper 1 and paper 2 in each of Programming and Systems Analysis. In order to obtain a certain a candidate must pass both paper 1 examinations and at least one paper 2. If he passes all four papers he obtains a merit award. If he passes both paper 1 examinations but does not pass either paper 2 he is allowed a re-sit. In all other cases, he is given a failing grade. The results obtained by each candidate are punched in the first four columns of a data card, using the abbreviations:
P – pass in a paper;
F – fail in a paper;
N – paper not taken.
This program inputs the results obtained by a student and classifies his grade as one of merit, certificate, re-sit, or fail.