Create DictionaryOrder.pas
This program reads two words, each of up to four letters, and determines whether they are in dictionary order.
In this program, we first declare two string variables word1 and word2 to store the input words. We then prompt the user to enter the first and second word using write and readln statements.
We then use an if-else statement to compare the two words. If word1 is less than word2, we output a message saying that the words are in dictionary order. If the two words are equal, we output a message saying that the words are the same. If word1 is greater than word2, we output a message saying that the words are not in dictionary order.

Note that the comparison between the two words is done using the < operator, which compares the ASCII values of the characters in the two strings. This means that uppercase letters come before lowercase letters in the ASCII table, so if the user enters a mix of uppercase and lowercase letters, the program will treat them as if they are all uppercase