Create CheckLowercaseLetter.pas
This Pascal program that uses the Boolean expression letter := (character >= 'a') and (character <= 'z') to check if a given character is a lowercase letter:
In this program, we first declare two variables: character, which stores the character we want to check, and letter, which is a Boolean variable that will be assigned true if the character is a lowercase letter and false otherwise.

We then use the readln function to read a character from the user and assign it to the character variable.

Next, we use the Boolean expression letter := (character >= 'a') and (character <= 'z') to check if the character is a lowercase letter. If the character is a lowercase letter, letter is assigned true. Otherwise, letter is assigned false.

Finally, we use an if-else statement to output a message to the console indicating whether the character is a lowercase letter or not. We use the writeln function to output the appropriate message to the console.